+++ R/rd-describe-in.R
@@ -23,58 +23,98 @@ topic_add_describe_in <- function(topic, block, env) {
   }
 
   dest <- find_object(tag$val$name, env)
-  label <- build_label(block$object, dest, block)
-  if (is.null(label))
-    return()
+  metadata <- build_minidesc_metadata(block$object, dest)
 
   topic$add(rd_section_minidesc(
-    label$type,
-    label$label,
-    tag$val$description
+    name = block$object$topic,
+    desc = tag$val$description,
+    extends = metadata$extends,
+    generic = metadata$generic,
+    class = metadata$class
   ))
   dest$topic
 }
 
 # Field -------------------------------------------------------------------
 
-rd_section_minidesc <- function(type, label, desc) {
-  stopifnot(is.character(type), is.character(label), is.character(desc))
-  stopifnot(length(desc) == length(label))
+#' Record data for minidescription sections from `@describeIn`
+#'
+#' @param name name of the source function.
+#' @param desc description passed to `@describeIn`.
+#' @param extends how the source function extends the destination function:
+#' - `"generic"` if the source extends a (S3 or S4) generic in the destination,
+#' - `"class"` if the source extends an informal S3 or formal S4 constructor
+#'    in the destination.
+#'    For S3, there is always only *one* class.
+#'    For S4, the methods' signature is used instead, to cover multiple dispatch.
+#' - `""` (default) otherwise.
+#' @param generic,class name of the generic and class that is being extended by
+#' the method, otherwise empty string (`""`).
+#' @return a dataframe with one row for each `@describeIn`, wrapped inside
+#' `rd_section()`
+#' @noRd
+rd_section_minidesc <- function(name,
+                                desc,
+                                extends = c("", "generic", "class"),
+                                generic = "",
+                                class = "") {
+  stopifnot(is_string(name))
+  stopifnot(is_character(desc))
+  rlang::arg_match(extends)
+  stopifnot(is_string(generic))
+  stopifnot(is_string(class))
 
-  rd_section("minidesc", list(type = type, desc = desc, label = label))
+  data <- data.frame(
+    name = name,
+    desc = desc,
+    extends = extends,
+    generic = generic,
+    class = class,
+    stringsAsFactors = FALSE
+  )
+  rd_section("minidesc", data)
 }
 
 #' @export
 merge.rd_section_minidesc <- function(x, y, ..., block) {
   stopifnot(identical(class(x), class(y)))
 
-  if (!identical(x$value$type, y$value$type)) {
-    warn_roxy_block(block, "Don't know how to combine @describeIn types {.str {x$value$type}} and {.str {y$value$type}}")
-    x
-  } else {
-    rd_section_minidesc(
-      x$value$type,
-      label = c(x$value$label, y$value$label),
-      desc = c(x$value$desc, y$value$desc)
-    )
-  }
+  rd_section("minidesc", rbind(x$value, y$value))
 }
+
+# Rd Output -------------------------------------------------------------------
 
 #' @export
 format.rd_section_minidesc <- function(x, ...) {
-  title <- switch(x$value$type,
-    generic = "Methods (by class)",
-    class = "Methods (by generic)",
-    "function" = "Functions"
+  order <- intersect(c("generic", "class", ""), unique(x$value$extends))
+  by <- factor(x$value$extends, levels = order)
+  subsections <- split(x$value, by)
+  body <- purrr::map2_chr(subsections, names(subsections), format_section)
+
+  paste0(body, collapse = "\n")
+}
+
+format_section <- function(df, type) {
+  if (type == "class") {
+    title <- paste0(
+      "Methods for class \\code{", escape(df$class[[1]]), "}"
+    )
+  } else if (type == "generic") {
+    title <- paste0(
+      "Methods for generic \\code{", escape(df$generic[[1]]), "()}"
+    )
+  } else {
+    title <- "Related functions"
+  }
+
+  bullets <- paste0("\\code{", escape(df$name), "}: ", df$desc)
+  body <- paste0(
+    "\\itemize{\n",
+    paste0("  \\item ", bullets, "\n", collapse = ""),
+    "}\n"
   )
 
-  paste0(
-    "\\section{", title, "}{\n",
-    "\\itemize{\n",
-    paste0("\\item \\code{", escape(x$value$label), "}: ", x$value$desc,
-      collapse = "\n\n"),
-    "\n}}\n"
-  )
+  paste0("\\section{", title, "}{\n", body, "}")
 }
 
 # Helpers -----------------------------------------------------------------
@@ -92,36 +132,68 @@ find_object <- function(name, env) {
   }
 }
 
-build_label <- function(src, dest, block) {
+#' Build metadata for how to present `@describeIn` tag
+#' @return list of character scalars named `extends`, `generic`, `class`.
+#' See rd_section_minidesc() for details.
+#' @noRd
+build_minidesc_metadata <- function(src, dest) {
   src_type <- class(src)[1]
   dest_type <- class(dest)[1]
+  dest_name <- as.character(dest$topic)
 
-  if (dest_type == "s4class" && src_type == "s4method") {
-    # Label S4 methods in class with their generic
-    type <- "class"
-    label <- as.character(src$value@generic)
-  } else if (dest_type == "s4generic" && src_type == "s4method") {
-    # Label S4 methods in generic with their signature
-    type <- "generic"
-    sig <- src$value@defined
-    if (length(sig) == 1) {
-      label <- as.character(sig)
+  if (src_type == "s3method") {
+    generic <- attr(src$value, "s3method")[1]
+    class <- attr(src$value, "s3method")[2]
+    if (dest_type == "s3generic" && generic == dest_name) {
+      # src method fits dest generic
+      extends <- "generic"
+    } else if (fits_constructor(dest_name, src)) {
+      # src method fits informal dest constructor (heuristically)
+      extends <- "class"
     } else {
-      label <- paste0(names(sig), " = ", sig, collapse = ",")
+      extends <- ""
     }
-  } else if (dest_type == "function" && src_type == "s3method") {
-    # Assuming you document S3 methods in the class constructor
-    type <- "class"
-    label <- attr(src$value, "s3method")[1]
-  } else if (dest_type == "s3generic" && src_type == "s3method") {
-    # Label S3 methods in generic with their class
-    type <- "generic"
-    label <- attr(src$value, "s3method")[2]
+  } else if (src_type == "s4method") {
+    generic <- as.character(src$value@generic)
+    class <- sig2class(src$value@defined)
+    if (dest_type == "s4generic") {
+      # TODO must test whether src method fits dest generic
+      extends <- "generic"
+    } else if (dest_type == "s4class") {
+      extends <- "class"
+      # TODO must test whether src method fits dest constructor
+    } else {
+      extends <- ""
+    }
   } else {
-    # Otherwise just fallback to function + topic
-    type <- "function"
-    label <- src$topic
+    generic <- ""
+    class <- ""
+    extends <- ""
+  }
+  list(extends = extends, generic = generic, class = class)
+}
+
+# Turn S4 signature into a string
+sig2class <- function(sig) {
+  if (length(sig) == 1) {
+    as.character(sig)
+  } else {
+    paste0(names(sig), " = ", sig, collapse = ",")
+  }
+}
+
+# Is destination is probably constructor for src?
+fits_constructor <- function(dest_name, src) {
+  src_class <- attr(src$value, "s3method")[2]
+
+  # simple case where class name is the same as the constructor name
+  if (src_class == dest_name) {
+    return(TRUE)
   }
 
-  list(type = type, label = label)
+  # more complex case where class name = package name + constructor name
+  evalenv <- roxy_meta_get("env") %||% parent.frame() # needed for tests
+  pkg_name <- utils::packageName(evalenv) %||% ""
+
+  src_class == paste0(pkg_name, "_", dest_name)
 }