+++ NEWS.md
@@ -1,5 +1,11 @@
 # roxygen2 (development version)
 
+* `@describeIn` can now be used to combine more types of functions
+  (generics, methods and other functions) into a single topic.
+  The resulting section organises the functions by type (#1181).
+  Methods are recognized only if they extend the generic in the destination,
+  or if the destination can heuristically be identified as a constructor.
+
 * All built-in tags are now documented so that you can do (e.g.) `?"@param"`
   to get a basic description of `@param` and a pointer where to learn more 
   (#1165).
@@ -41,8 +47,6 @@
   documentation (#1354).
 
 * Automated usage no longer mangles nbsp in default arguments (#1342).
-
-* Give useful error if `@describeIn` can't combine docs (#1366).
 
 * Code evaluated in inline markdown code chunks and `@eval`/`@evalRd`/
   `@evalNamespace` is now evaluated in an environment designed to be more